+++ src/habit/habit.service.ts
@@ -16,10 +16,10 @@ export class HabitService {
     this.database = await this.databaseService.getDatabase();
   }
 
-  async getHabitById(id: number): Promise<HabitDto> {
+  async getHabitById(userId: number, id: number): Promise<HabitDto> {
     const result = await this.database.query(
-      'SELECT * FROM habits WHERE id = $1',
-      [id],
+      'SELECT * FROM habits WHERE id = $1 AND user_id = $2',
+      [id, userId],
     );
 
     if (result.rows.length > 0) {
@@ -29,44 +29,52 @@ export class HabitService {
     }
   }
 
-  async getAllHabits(): Promise<HabitDto[]> {
-    const result = await this.database.query('SELECT * FROM habits');
-    return result.rows;
-  }
-
-  async getLatestHabit(): Promise<HabitDto[]> {
+  async getAllHabits(userId: number): Promise<HabitDto[]> {
     const result = await this.database.query(
-      'SELECT * FROM habits WHERE habits.deleted_at IS NULL ORDER by habits.created_at ASC',
+      'SELECT * FROM habits WHERE user_id = $1',
+      [userId],
     );
     return result.rows;
   }
 
-  async getAllHabitsGroupedByDate(): Promise<HabitDto[]> {
-    const result = await this.database.query(`
-          SELECT DATE(given_at) AS date,
-          JSON_AGG(json_build_object('id', id, 'text', text, 'given_at', given_at, 
+  async getLatestHabit(userId: number): Promise<HabitDto[]> {
+    const result = await this.database.query(
+      'SELECT * FROM habits WHERE user_id = $1 AND habits.deleted_at IS NULL ORDER by habits.created_at ASC',
+      [userId],
+    );
+    return result.rows;
+  }
+
+  async getAllHabitsGroupedByDate(userId: number): Promise<HabitDto[]> {
+    const result = await this.database.query(
+      `
+          SELECT DATE(created_at) AS date,
+          JSON_AGG(json_build_object('id', id, 'title', title, 'type', type, 'repeat', repeat,
           'created_at', created_at, 'updated_at', updated_at, 
-          'deleted_at', deleted_at) ORDER BY given_at DESC) AS habits
+          'deleted_at', deleted_at) ORDER BY created_at DESC) AS habits
           FROM habits
-          WHERE deleted_at IS NULL
+          WHERE user_id = $1 AND deleted_at IS NULL
           GROUP BY date
           ORDER BY date DESC;
-        `);
-    return result.rows;
-  }
-
-  async getDailyHabits(): Promise<HabitDto[]> {
-    const result = await this.database.query(
-      `SELECT * FROM habits WHERE habits.repeat = 'daily'`,
+        `,
+      [userId],
     );
     return result.rows;
   }
 
-  async insertHabit(habit: HabitDto) {
+  async getDailyHabits(userId: number): Promise<HabitDto[]> {
+    const result = await this.database.query(
+      `SELECT * FROM habits WHERE user_id = $1 AND habits.repeat = 'daily'`,
+      [userId],
+    );
+    return result.rows;
+  }
+
+  async insertHabit(userId: number, habit: HabitDto) {
     try {
       const result = await this.database.query(
-        'INSERT INTO habits(title, type, repeat) VALUES($1, $2, $3) RETURNING *',
-        [habit.title, habit.type, habit.repeat],
+        'INSERT INTO habits(title, type, repeat, user_id) VALUES($1, $2, $3, $4) RETURNING *',
+        [habit.title, habit.type, habit.repeat, userId],
       );
 
       console.log('Habit inserted successfully:', result.rows[0]);
@@ -80,16 +88,17 @@ export class HabitService {
     }
   }
 
-  async updateHabit(id: number, updatedHabit: HabitDto) {
+  async updateHabit(userId: number, id: number, updatedHabit: HabitDto) {
     try {
       const result = await this.database.query(
-        'UPDATE habits SET title = $1, type = $2, deleted_at = $3, updated_at = $4 WHERE id = $5 RETURNING *',
+        'UPDATE habits SET title = $1, type = $2, deleted_at = $3, updated_at = $4 WHERE id = $5 AND user_id = $6 RETURNING *',
         [
           updatedHabit.title,
           updatedHabit.type,
           updatedHabit.deleted_at,
           new Date().toISOString(),
           id,
+          userId,
         ],
       );
 
@@ -108,11 +117,11 @@ export class HabitService {
     }
   }
 
-  async deleteHabit(id: number) {
+  async deleteHabit(userId: number, id: number) {
     try {
       const result = await this.database.query(
-        'DELETE FROM habits WHERE id = $1 RETURNING *',
-        [id],
+        'DELETE FROM habits WHERE id = $1 AND user_id = $2 RETURNING *',
+        [id, userId],
       );
 
       if (result.rows.length > 0) {
@@ -130,12 +139,12 @@ export class HabitService {
     }
   }
 
-  async removeHabit(id: number) {
+  async removeHabit(userId: number, id: number) {
     const now = new Date().toISOString();
     try {
       const result = await this.database.query(
-        'UPDATE habits SET deleted_at = $1 WHERE id = $2 RETURNING *',
-        [now, id],
+        'UPDATE habits SET deleted_at = $1 WHERE id = $2 AND user_id = $3 RETURNING *',
+        [now, id, userId],
       );
 
       if (result.rows.length > 0) {
